+++ docs/Whats-New.md
@@ -370,6 +370,8 @@ New:
 - Customize airstrike targets (by NetsuNegi)
 - Aggressive attack move mission (by CrimRecya)
 - Amphibious access vehicle (by CrimRecya)
+- Attached effect attach/discard by health (by Ollerus)
+- Attached effect with `ExtraWarheads` (by Ollerus)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)